Add basic stale pid detection heuristic
While any normal crashes should already remove the pid file
if one was created mumo can still be killed in ways which
make it unable to clean up after itself. This patch adds
a basic heuristic which checks if the pid mentioned in the
pid file belongs to a running process. If so we assume this
is actually mumo and terminate. If not we assume mumo isn't
running and break the lock. This approach is not 100% reliable
and it break the acquire timeout properties we had before
but should do the trick.